# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: TogoMarket API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /admin/events:
    post:
      operationId: adminCreateEvent
      tags: [admin, events]
      summary: Create a new event (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminCreateEventInput"
      responses:
        "201":
          description: Event created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
        "403":
          description: Forbidden
  /admin/events/delete:
    post:
      operationId: adminDeleteEvent
      tags: [admin, events]
      summary: Delete an event (admin only)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/AdminDeleteInput"
      responses:
        "200":
          description: Deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DeleteResult"
        "403":
          description: Forbidden
...
  /contact-requests:
    post:
      tags: [listings]
      operationId: createContactRequest
      summary: Record a buyer contact unlock request
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ContactRequestInput'
      responses:
        '200':
          description: Request recorded.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SuccessResult'
        '404':
          description: Listing not found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
        '500':
          description: Internal error.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /vendor/contact-requests:
    post:
      tags: [vendors]
      operationId: getVendorContactRequests
      summary: Get contact request stats for a vendor's listings
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VendorLoginInput'
      responses:
        '200':
          description: Stats per listing.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ContactRequestStat'
        '403':
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
  /admin/contact-requests/stats:
    post:
      tags: [admin]
      operationId: getAdminContactStats
      summary: Get all contact request stats (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/AdminVerifyInput'
      responses:
        '200':
          description: Stats per listing.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AdminContactStat'
        '403':
          description: Forbidden.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorEnvelope'
components:
  schemas:
...
    Event:
      type: object
      required: [id, title, description, date, location, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        description:
          type: string
        flyerImage:
          type: ["string", "null"]
        videoPath:
          type: ["string", "null"]
        date:
          type: string
        location:
          type: string
        ticketLink:
          type: ["string", "null"]
        ticketPrice:
          type: ["string", "null"]
        createdAt:
          type: string
        isPublished:
          type: ["boolean", "null"]
    AdminCreateEventInput:
      type: object
      required: [password, title, description, date, location]
      properties:
        password:
          type: string
        title:
          type: string
        description:
          type: string
        flyerImage:
          type: string
        videoPath:
          type: string
        date:
          type: string
        location:
          type: string
        ticketLink:
          type: string
        ticketPrice:
          type: string
...
    DeleteResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
...
    AdminDeleteInput:
      type: object
      required: [id, password]
      properties:
        id:
          type: integer
        password:
          type: string
...
    AdminVerifyInput:
      type: object
      required: [password]
      properties:
        password:
          type: string
...
    VendorLoginInput:
      type: object
      required: [phone, password]
      properties:
        phone:
          type: string
        password:
          type: string
...
    SuccessResult:
      type: object
      required: [success]
      properties:
        success:
          type: boolean
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string
    ContactRequestInput:
      type: object
      required: [listingId, buyerName, buyerPhone]
      properties:
        listingId:
          type: integer
        buyerName:
          type: string
        buyerPhone:
          type: string
    ContactRequestStat:
      type: object
      required: [listingId, listingName, count]
      properties:
        listingId:
          type: integer
        listingName:
          type: string
        count:
          type: integer
    AdminContactStat:
      type: object
      required: [listingId, listingName, vendorPhone, count]
      properties:
        listingId:
          type: integer
        listingName:
          type: string
        vendorPhone:
          type: string
        count:
          type: integer
...